% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdist.R
\name{rdist}
\alias{rdist}
\title{simulation d´une loi discret}
\usage{
rdist(x, p)
}
\arguments{
\item{x}{numeric vector representing the value of the random variable}

\item{p}{numeric vector representing probabilities}
}
\description{
simulation d´une loi discret
}
