% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdistk.R
\name{rdistk}
\alias{rdistk}
\title{simulation d´une discret k fois}
\usage{
rdistk(x, p, k)
}
\arguments{
\item{x}{numeric vector representing the value of the random variable}

\item{p}{numeric vector representing probabilities}

\item{k}{numeric representing the number of repetition}
}
\description{
simulation d´une discret k fois
}
