% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimCMk.R
\name{SimCMk}
\alias{SimCMk}
\title{simulation d´une chaine de markov a espace des etats fini}
\usage{
SimCMk(P, mu, n)
}
\arguments{
\item{P}{matrix of transition}

\item{mu}{numeric vector of initial distrebution}

\item{n}{numeric number of repetition}
}
\description{
simulation d´une chaine de markov a espace des etats fini
}
